MC3/pyproject.toml
[tool.poetry]
name = "mc3"
version = "0.1.0"
description = ""
authors = ["Averniko <averniko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.108.0"}
aiokafka = "^0.10.0"
httpx = "^0.26.0"
opentelemetry-instrumentation-fastapi = "^0.45b0"
opentelemetry-instrumentation-logging = "^0.45b0"
opentelemetry-exporter-otlp = "^1.24.0"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-bugbear = "^23.12.2"
flake8-builtins = "^2.2.0"
flake8-commas = "^2.1.0"
flake8-variables-names = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


